Small bug: spatial domain discretization is wrong
This is a small and commont bug when using FFTs. The domain on each axis is (0, 2π). When discretizing it, the code was including both ends! What this means is that actually the domain has a period of 2π * (1 + 1/(N-1)). This introduces noise and aliasing. And the periodic function is discontinuous.

The domain should be divided in `N_POINTS_P_AXIS` segments and the points should correspond to the left point of each segment.

It is difficult to see but this appears to reduce noise in the solution.