Move contributions to subfolder
There are many great contributions from the community that extend some
of the scripts that are written for the YouTube videos. Unfortunately, I
cannot validate their implementation, but I will leave them since they
can still be of great help for someone trying to find an implementation
for a more advanced concept.